requirements.txt
anyjson==0.3.3
arrow==0.10.0
astroid==1.6.5
atomicwrites==1.2.1
attrs==18.2.0
bootstrap-admin==0.4.0
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.1
defusedxml==0.5.0
Django==2.1.2
django-admin-bootstrapped==2.5.7
django-admin-generator==2.1.0
django-allauth==0.38.0
django-crispy-forms==1.7.0
django-fcm==0.1.1
django-flat-responsive==2.0
django-mysql==2.4.1
django-rest-auth==0.9.3
django-templated-mail==1.0.0
django-tinymce==2.6.0
django-utils2==2.5.0
django-widget-tweaks==1.4.2
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
djet==0.2.1
djoser==1.0.1
fcm-django==0.2.11
flake8==3.5.0
idna==2.6
include==0.2.2
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.6.0
pycodestyle==2.3.1
pyfcm==1.3.1
pyflakes==1.6.0
PyJWT==1.5.3
pylint==1.9.3
pytest==3.6.2
pytest-cov==2.5.1
pytest-django==3.3.2
pytest-pythonpath==0.7.2
python-dateutil==2.6.1
python-utils==2.3.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rest-social-auth==1.2.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.5.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
